CIRCSTORE-426 move event handling to separate handler
