+++ src/main/java/org/folio/EventConsumerVerticle.java
@@ -4,6 +4,7 @@ import static java.util.stream.Collectors.toList;
 import static org.folio.rest.tools.utils.ModuleName.getModuleName;
 import static org.folio.rest.tools.utils.ModuleName.getModuleVersion;
 import static org.folio.service.event.InventoryEventType.INVENTORY_ITEM_UPDATED;
+import static org.folio.service.event.InventoryEventType.INVENTORY_PICKUP_LOCATION_UPDATED;
 import static org.folio.service.event.InventoryEventType.INVENTORY_SERVICE_POINT_UPDATED;
 import static org.folio.support.kafka.KafkaConfigConstants.KAFKA_ENV;
 import static org.folio.support.kafka.KafkaConfigConstants.KAFKA_HOST;
@@ -26,6 +27,7 @@ import org.folio.kafka.services.KafkaTopic;
 import org.folio.service.event.InventoryEventType;
 import org.folio.service.event.handler.ItemUpdateEventHandler;
 import org.folio.service.event.handler.ServicePointUpdateEventHandler;
+import org.folio.service.event.handler.UpdatePickupLocationEventHandler;
 
 import io.vertx.core.AbstractVerticle;
 import io.vertx.core.CompositeFuture;
@@ -78,6 +80,8 @@ public class EventConsumerVerticle extends AbstractVerticle {
     return createInventoryEventConsumer(INVENTORY_ITEM_UPDATED, config, new ItemUpdateEventHandler(context))
       .compose(r -> createInventoryEventConsumer(INVENTORY_SERVICE_POINT_UPDATED, config,
         new ServicePointUpdateEventHandler(context)))
+      .compose(r -> createInventoryEventConsumer(INVENTORY_PICKUP_LOCATION_UPDATED, config,
+        new UpdatePickupLocationEventHandler(context)))
       .mapEmpty();
   }
