+++ src/main/java/org/folio/persist/RequestPolicyRepository.java
@@ -1,13 +1,17 @@
 package org.folio.persist;
 
+import static org.folio.rest.persist.PgUtil.postgresClient;
 import static org.folio.support.ModuleConstants.REQUEST_POLICY_CLASS;
 import static org.folio.support.ModuleConstants.REQUEST_POLICY_TABLE;
 
+import java.util.Map;
+
 import org.folio.rest.jaxrs.model.RequestPolicy;
-import org.folio.rest.persist.PostgresClient;
+
+import io.vertx.core.Context;
 
 public class RequestPolicyRepository extends AbstractRepository<RequestPolicy> {
-  public RequestPolicyRepository(PostgresClient postgresClient) {
-    super(postgresClient, REQUEST_POLICY_TABLE, REQUEST_POLICY_CLASS);
+  public RequestPolicyRepository(Context context, Map<String, String> okapiHeaders) {
+    super(postgresClient(context, okapiHeaders), REQUEST_POLICY_TABLE, REQUEST_POLICY_CLASS);
   }
 }
